pepper/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar


Flask == 1.0.2
Flask_Bcrypt == 0.7.1
Flask_Login == 0.4.1
Flask_SQLAlchemy == 2.4.0
Pillow == 6.0.0
itsdangerous == 0.24
Flask-Cors==3.0.6
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-WTF==0.14.2
Flask-Login==0.4.1
WTForms==2.2.1
